move releasenote check to the linter namespace
