+++ tasks/linter.py
@@ -1,3 +1,4 @@
+import os
 import re
 import sys
 from collections import defaultdict
@@ -16,7 +17,8 @@ from tasks.libs.ciproviders.gitlab import (
     load_context,
 )
 from tasks.libs.common.check_tools_version import check_tools_version
-from tasks.libs.common.utils import DEFAULT_BRANCH, color_message
+from tasks.libs.ciproviders.github_api import GithubAPI
+from tasks.libs.common.utils import DEFAULT_BRANCH, GITHUB_REPO_NAME, color_message
 from tasks.libs.types.copyright import CopyrightLinter
 from tasks.modules import GoModule
 from tasks.test_core import ModuleLintResult, process_input_args, process_module_results, test_core
@@ -393,3 +395,39 @@ def gitlab_ci(_, test="all", custom_context=None):
         if not res["valid"]:
             print(color_message(f"Errors: {res['errors']}", "red"), file=sys.stderr)
             raise Exit(code=1)
+
+
+def is_pr_context(branch, pr_url, test_name):
+    if branch == DEFAULT_BRANCH:
+        print(f"Running on {DEFAULT_BRANCH}, skipping check for {test_name}.")
+        return False
+    if not pr_url:
+        print(f"PR not found, skipping check for {test_name}.")
+        return False
+    return True
+
+
+@task
+def releasenote(ctx):
+    """
+    Lint release notes with Reno
+    """
+    branch = os.environ.get("CIRCLE_BRANCH")
+    pr_url = os.environ.get("CIRCLE_PULL_REQUEST")
+
+    run_check = is_pr_context(branch, pr_url, "team assignment")
+    if run_check:
+        github = GithubAPI(repository=GITHUB_REPO_NAME, public_repo=True)
+        pr_id = pr_url.rsplit('/')[-1]
+        if github.is_release_note_needed(pr_id):
+            if not github.contains_release_note(pr_id):
+                print(
+                    "Error: No releasenote was found for this PR. Please add one using 'reno'"
+                    ", see https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno"
+                    ", or apply the label 'changelog/no-changelog' to the PR."
+                )
+                raise Exit(code=1)
+        else:
+            print("'changelog/no-changelog' label found on the PR: skipping linting")
+
+    ctx.run("reno lint")
